// src/Application/Jobs/Commands/CreateJob/CreateJobCommand.d.ts
export default class CreateJobCommand {
    constructor(Cnpj: string, RazaoSocial: string, EmailRecrutador: string, NomeRecrutador: string, CelularRecrutador: string, TipoVaga: number, // 
    Cargo: string, NumeroVagas: number, TipoEstabelecimento: string, DescricaoEmpresa: string, FormaTrabalho: number, Cidade: string, Bairro: string, DescricaoCargo: string, FormacaoExperiencia: number, CaracteristicasComportamentais: string, FormaRemuneracao: number, Beneficios: string, JornadaTrabalhoCargo: number, FormaRecebimentoCurriculo: number, EmailRecebimentoCurriculo: string);
    Cnpj: string;
    RazaoSocial: string;
    EmailRecrutador: string;
    NomeRecrutador: string;
    CelularRecrutador: string;
    TipoVaga: number;
    Cargo: string;
    NumeroVagas: number;
    TipoEstabelecimento: string;
    DescricaoEmpresa: string;
    FormaTrabalho: number;
    Cidade: string;
    Bairro: string;
    DescricaoCargo: string;
    FormacaoExperiencia: number;
    CaracteristicasComportamentais: string;
    FormaRemuneracao: number;
    Beneficios: string;
    JornadaTrabalhoCargo: number;
    FormaRecebimentoCurriculo: number;
    EmailRecebimentoCurriculo: string;
}
//# sourceMappingURL=CreateJobCommand.d.ts.map